Move WKWEbViewConfiguration extension to it's own file
